Helper functions and comments
